Reorganize client sample, remove EventBus
