+++ samples/client/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <service android:name="global.covesa.sdk.client.PushServiceImpl"
+        <service android:name="global.covesa.sdk.client.push.PushServiceImpl"
             android:exported="false">
             <intent-filter>
                 <action android:name="global.covesa.sdk.PUSH_EVENT"/>
