+++ samples/client/src/main/java/global/covesa/sdk/client/MainActivity.kt
@@ -6,10 +6,12 @@ import androidx.activity.compose.setContent
 import androidx.lifecycle.viewmodel.compose.viewModel
 import global.covesa.sdk.api.client.LightsServiceClient
 import global.covesa.sdk.api.client.ServicesCatalogClient
+import global.covesa.sdk.client.push.ActionEvent
 import global.covesa.sdk.client.ui.MainUi
 import global.covesa.sdk.client.ui.theme.CovesaSDKTheme
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
 
 
@@ -36,7 +38,9 @@ class MainActivity : ComponentActivity() {
 
     private fun subscribeActions() {
         CoroutineScope(Dispatchers.IO).launch {
-            EventBus.subscribe<ActionEvent> { it.handleAction(this@MainActivity) }
+            ActionEvent.events.collect {
+                it.handleAction(this@MainActivity)
+            }
         }
     }
 }
