--- samples/client/src/main/java/global/covesa/sdk/client/PushSubscriptionEvent.kt
@@ -1,5 +1,0 @@
-package global.covesa.sdk.client
-
-data class PushSubscriptionEvent(
-    val registered: Boolean
-)
